batcheval: catch pebble data corruption on split/merge commit triggers
This commit catches the pebble data corruption error on split and merge
commit triggers (instead of crashing the node). This is useful if a
file that an external SSTable is pointing to gets deleted, we don't want
to crash the process. Excise command could be used to recover from this
situation.

References: #143135

Release note: None